# api/src/config/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PulseOps V2 API",
    "version": "2.0.0",
    "description": "Enterprise platform API for PulseOps V2. Supports database management, authentication (JWT + HttpOnly cookies), module hot-dropping, and system configuration. All URLs are defined in urls.json — this spec references relative paths under the /api prefix.",
    "contact": {
      "name": "PulseOps Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Relative to host — works on any environment"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained via POST /auth/login. Expiry: 24h."
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "HttpOnly cookie set automatically on login. Same JWT as BearerAuth."
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "description": "Standard success envelope returned by all endpoints on success.",
        "properties": {
          "success": { "type": "boolean", "example": true, "description": "Always true on success" },
          "data": { "type": "object", "description": "Endpoint-specific response payload" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error envelope returned by all endpoints on failure.",
        "properties": {
          "success": { "type": "boolean", "example": false, "description": "Always false on error" },
          "error": {
            "type": "object",
            "properties": {
              "message": { "type": "string", "description": "Human-readable error message from APIErrors.json", "example": "Invalid credentials" },
              "code": { "type": "string", "description": "Machine-readable error code", "example": "TOKEN_EXPIRED" }
            }
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "description": "Credentials for user authentication.",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "description": "User email address", "example": "admin@test.com" },
          "password": { "type": "string", "format": "password", "description": "User password (min 8 chars)", "example": "Infosys@123" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "description": "Returned on successful login. Access token also set as HttpOnly cookie.",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "accessToken": { "type": "string", "description": "JWT access token (24h expiry)", "example": "eyJhbGciOiJIUzI1NiIs..." },
              "refreshToken": { "type": "string", "description": "JWT refresh token (7d expiry)", "example": "eyJhbGciOiJIUzI1NiIs..." },
              "expiresIn": { "type": "integer", "description": "Access token TTL in seconds", "example": 86400 },
              "user": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "admin@test.com" },
                  "name": { "type": "string", "example": "Default Admin" },
                  "role": { "type": "string", "enum": ["super_admin", "admin", "user"], "example": "super_admin" }
                }
              }
            }
          }
        }
      },
      "UserProfile": {
        "type": "object",
        "description": "Current authenticated user profile.",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "id": { "type": "integer", "description": "User ID (database provider only)", "example": 1 },
              "email": { "type": "string", "example": "admin@test.com" },
              "name": { "type": "string", "example": "Default Admin" },
              "role": { "type": "string", "example": "super_admin" },
              "status": { "type": "string", "enum": ["active", "inactive"], "example": "active" },
              "last_login": { "type": "string", "format": "date-time", "nullable": true }
            }
          }
        }
      },
...
      "ConfigEntry": {
        "type": "object",
        "description": "A key-value configuration entry stored as JSONB.",
        "required": ["key", "value"],
        "properties": {
          "key": { "type": "string", "description": "Unique config key", "example": "auth_provider" },
          "value": { "type": "object", "description": "JSON value (any structure)", "example": { "provider": "json_file" } },
          "description": { "type": "string", "description": "Human-readable description", "example": "Active authentication provider" }
        }
      },
      "ConfigListResult": {
        "type": "object",
        "description": "List of all configuration entries.",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "key": { "type": "string", "example": "auth_provider" },
                "value": { "type": "object" },
                "description": { "type": "string" },
                "updated_at": { "type": "string", "format": "date-time" }
              }
            }
          }
        }
      },
      "AuthProviderRequest": {
        "type": "object",
        "description": "Request to change the active authentication provider.",
        "required": ["provider"],
        "properties": {
          "provider": { "type": "string", "enum": ["json_file", "database", "social"], "description": "Target provider ID", "example": "database" }
        }
      },
      "AuthProviderConfigResult": {
        "type": "object",
        "description": "Current authentication provider configuration.",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "activeProvider": { "type": "string", "example": "json_file" },
              "availableProviders": {
                "type": "array",
                "items": { "type": "string" },
                "example": ["json_file", "database", "social"]
              }
            }
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "description": "Request body for token refresh. Send refreshToken from login response.",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": { "type": "string", "description": "JWT refresh token from login", "example": "eyJhbGciOiJIUzI1NiIs..." }
        }
      }
    }
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "login",
        "summary": "Login with email and password",
        "description": "Authenticates the user against the active provider (json_file reads from DefaultAdminUser.json, database reads from system_users table with bcrypt). On success, returns JWT access token (24h) and refresh token (7d), and sets HttpOnly cookies (accessToken, refreshToken) for browser-based auth.",
        "requestBody": {
          "required": true,
          "description": "User credentials. Both fields are required.",
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful. Tokens returned in body and set as HttpOnly cookies.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginResponse" } } }
          },
          "400": {
            "description": "Missing email or password in request body.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Invalid email or password.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Account exists but is inactive.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/auth/config": {
      "get": {
        "tags": ["Authentication"],
        "operationId": "getAuthConfig",
        "summary": "Get auth provider config",
        "description": "Returns the active authentication provider and list of all available providers. Public endpoint — no auth required.",
        "responses": {
          "200": {
            "description": "Auth provider configuration.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthProviderConfigResult" } } }
          }
        }
      },
      "post": {
        "tags": ["Authentication"],
        "operationId": "setAuthConfig",
        "summary": "Set auth provider",
        "description": "Changes the active authentication provider. Requires super_admin role. If switching to 'database', validates that the database is initialized and has default data. Writes to auth-provider.json and system_config table.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "description": "Target provider. Must be one of: json_file, database, social.",
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthProviderRequest" } } }
        },
        "responses": {
          "200": { "description": "Provider updated successfully.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "400": { "description": "Invalid provider ID or database not initialized.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "No valid JWT token provided.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Authenticated user lacks super_admin role.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "operationId": "getMe",
        "summary": "Get current user profile",
        "description": "Returns the profile of the currently authenticated user. For database provider, fetches fresh data from system_users. For json_file provider, returns the decoded JWT payload.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserProfile" } } } },
          "401": { "description": "Not authenticated or token expired.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "User not found in database.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "refreshToken",
        "summary": "Refresh access token",
        "description": "Exchanges a valid refresh token for a new access token. Also sets updated HttpOnly cookies. Use this when the access token expires (24h) but the refresh token is still valid (7d).",
        "requestBody": {
          "required": true,
          "description": "The refresh token received from the login response.",
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshTokenRequest" } } }
        },
        "responses": {
          "200": {
            "description": "New access token generated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": { "type": "string" },
                        "expiresIn": { "type": "integer", "example": 86400 }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Refresh token is invalid or expired.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "logout",
        "summary": "Logout",
        "description": "Clears HttpOnly cookies (accessToken, refreshToken) from the browser. The JWT itself is not invalidated server-side (stateless). Client should also clear localStorage.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Logged out. Cookies cleared.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": { "type": "string", "example": "Logged out successfully" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/config": {
      "get": {
        "tags": ["Configuration"],
        "operationId": "listConfig",
        "summary": "List all config entries",
        "description": "Returns all key-value configuration entries from the system_config table, ordered by key. Requires JWT authentication.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "All configuration entries.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfigListResult" } } } },
          "401": { "description": "Authentication required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "post": {
        "tags": ["Configuration"],
        "operationId": "upsertConfig",
        "summary": "Create or update config entry",
        "description": "Inserts a new config entry or updates an existing one (upsert on key). Value is stored as JSONB. Requires super_admin role.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "description": "Config entry to create or update.",
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ConfigEntry" } } }
        },
        "responses": {
          "200": { "description": "Config saved.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "400": { "description": "Missing key or value.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Authentication required.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Requires super_admin role.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  }
}